ClassicTypeManager
SimpleTypeManager renamed to ClassicTypeManager
ClassicTypeManager uses ClassicTypes
