+++ src/test/java/org/jpokemon/api/PokemonTypeTest.java
@@ -15,31 +15,30 @@ import org.jpokemon.api.classic.ClassicTypes;
  * Tests type properties and the use of ClassicTypes.
  * 
  * @author atheriel@gmail.com
- * @author Zack Taylor
+ * @author Zach Taylor
  */
 @RunWith(JUnit4.class)
 public class PokemonTypeTest {
 
 	@BeforeClass
 	public static void setup() {
-		// Instantiate and link all classic types with SimpleTypeManager
-		SimpleTypeManager.init();
-		ClassicTypes.init();
+		// Setup the hard-coded classic types
+		ClassicTypeManager.init();
 	}
 
 	/** Tests some properties of classic types. */
 	@Test
 	public void testClassicTypes() {
 		assertTrue("Dark is known.", PokemonType.manager.isRegistered(ClassicTypes.DARK));
-		assertTrue("Normal is ineffective against Ghost.", PokemonType.manager.getTypeByName("Normal").isIneffectiveAgainst("Ghost"));
-		assertFalse("Normal is not super-effective against Fighting.", PokemonType.manager.getTypeByName("Normal").isSuperEffectiveAgainst("Fighting"));
+		assertTrue("Normal is ineffective against Ghost.", PokemonType.manager.getByName("Normal").isIneffectiveAgainst("Ghost"));
+		assertFalse("Normal is not super-effective against Fighting.", PokemonType.manager.getByName("Normal").isSuperEffectiveAgainst("Fighting"));
 	}
 
 	/** Creates a new type and tests that it has been registered. */
 	@Test
 	public void testTypeAddition() {
 		PokemonType test = new PokemonType().setName("Test");
-		PokemonType.manager.registerType(test);
+		PokemonType.manager.register(test);
 		assertTrue("Test is now a type.", PokemonType.manager.isRegistered(test));
 	}
